Introducing golem.testutils.
Create new package golem.testutils for all unittest utility and fixture classes.
Move golem.tools.testdirfixture.TestDirFixture to golem.testutils.TempDirFixture.
